llama-indexer/requirements.txt
llama-index-agent-coa==0.2.0
llama-index-core==0.11.22
llama-index-embeddings-openai==0.2.5
llama-index-llms-openai==0.2.16
llama-index-llms-openai-like==0.2.0
llama-index-readers-file==0.2.2
llama-parse==0.5.13
openai==1.54.3
uvicorn==0.32.0
fastapi==0.115.4
requests==2.31.0